requirements.txt
Django==3.2.8
django_heroku==0.3.1
djangorestframework==3.12.4
djangorestframework_simplejwt==5.0.0
gunicorn==20.1.0
gensim~=4.1.2
docx2txt~=0.8